allennlp pipeline implementation according to pipeline interface
