+++ src/biome/text/pipelines/pipeline.py
@@ -1,4 +1,5 @@
 import copy
+import datetime
 import logging
 import os
 import pickle
@@ -9,10 +10,12 @@ from functools import lru_cache
 from logging.handlers import RotatingFileHandler
 from pathlib import Path
 from tempfile import mktemp
-from typing import cast, Type, Optional, List, Dict, Tuple, Any, Generic, TypeVar
+from typing import cast, Type, Optional, List, Dict, Tuple, Any, Generic, TypeVar, Union
+from gevent.pywsgi import WSGIServer
 
 import allennlp
 import numpy
+import pandas as pd
 import yaml
 from allennlp.common import JsonDict, Params
 from allennlp.common.checks import ConfigurationError
@@ -21,24 +24,33 @@ from allennlp.data import DatasetReader, Instance, Vocabulary
 from allennlp.data.dataset import Batch
 from allennlp.data.fields import LabelField
 from allennlp.data.token_indexers import SingleIdTokenIndexer
+from allennlp.interpret import SaliencyInterpreter
 from allennlp.models import Archive, Model
 from allennlp.modules import Embedding
 from allennlp.modules.seq2vec_encoders import PytorchSeq2VecWrapper
 from allennlp.modules.text_field_embedders import BasicTextFieldEmbedder
 from allennlp.predictors import Predictor
+from allennlp.service import server_simple
+
+from dask import dataframe as dd
+from dask_elk.client import DaskElasticClient
+from flask_cors import CORS
 from overrides import overrides
 from torch.nn import LSTM
 
+from biome.data import DataSource
+from biome.text import constants
+from biome.text.defs import TextClassifierPipeline, Pipeline
 from biome.text.pipelines._impl.allennlp.dataset_readers import DataSourceReader
 from biome.text.pipelines._impl.allennlp.learn import learn
 from biome.text.pipelines._impl.allennlp.models import (
     load_archive,
     SequenceClassifierBase,
 )
-from biome.text.pipelines.pipeline_interface import ClassifierPipeline
 from biome.text.pipelines._impl.allennlp.predictors.utils import (
     get_predictor_from_archive,
 )
+from biome.text.pipelines.defs import ExploreConfig, ElasticsearchConfig
 
 try:
     import ujson as json
@@ -61,7 +73,7 @@ Architecture = TypeVar("Architecture", bound=allennlp.models.Model)
 Reader = TypeVar("Reader", bound=DataSourceReader)
 
 
-class Pipeline(Generic[Architecture, Reader], Predictor, ClassifierPipeline):
+class Pipeline(Generic[Architecture, Reader], Predictor, TextClassifierPipeline):
     """
     This class combine the different allennlp components that make possible a ``Pipeline`,
     understanding as a model, not only the definition of the neural network architecture,
@@ -79,9 +91,6 @@ class Pipeline(Generic[Architecture, Reader], Predictor, ClassifierPipeline):
     reader
         The class:allennlp.data.DatasetReader
     """
-
-    def explore(self):
-        raise NotImplementedError
 
     _LOGGER = logging.getLogger(__name__)
 
@@ -228,7 +237,7 @@ class Pipeline(Generic[Architecture, Reader], Predictor, ClassifierPipeline):
         """
         return self._dataset_reader.signature
 
-    def predict(self, **inputs) -> dict:
+    def predict(self, *args, **inputs) -> dict:
         return self.predict_json(inputs)
 
     def _update_binary_path(self, path) -> None:
@@ -417,6 +426,208 @@ class Pipeline(Generic[Architecture, Reader], Predictor, ClassifierPipeline):
         decorated_func = lru_cache(maxsize=max_size)(self._predict_hashable_json)
 
         self.__setattr__("_predict_hashable_json", decorated_func)
+
+    def inputs_keys(self) -> List[str]:
+        return [k for k, v in self.signature.items() if not v.get("optional")]
+
+    def output(self) -> str:
+        return "label"
+
+    def explore(
+        self, ds_path: str, config: ExploreConfig, es_config: ElasticsearchConfig
+    ) -> dd.DataFrame:
+        from biome.text.pipelines._impl.allennlp.interpreters import (
+            IntegratedGradient as DefaultInterpreterClass,
+        )
+
+        def _interpret_dataframe(
+            df: pd.DataFrame,
+            pipeline: Pipeline,
+            interpreter_klass: Type = DefaultInterpreterClass,
+        ) -> pd.Series:
+            """
+            Apply a model interpretation to every partition dataframe
+
+            Parameters
+            ----------
+            df: pd.DataFrame
+                The partition DataFrame
+            pipeline: str
+                The pipeline
+            interpreter_klass: Type
+                The used interpreted class
+
+            Returns
+            -------
+
+            A pandas Series representing the interpretations
+
+            """
+
+            def interpret_row(
+                row: pd.Series, interpreter: SaliencyInterpreter
+            ) -> Union[dict, List[dict]]:
+                """Interpret a incoming dataframe row"""
+                data = row.to_dict()
+                interpretation = interpreter.saliency_interpret_from_json(data)
+                if len(interpretation) == 0:
+                    return {}
+                if len(interpretation) == 1:
+                    return interpretation[0]
+                return interpretation
+
+            interpreter = interpreter_klass(pipeline)
+            return df.apply(interpret_row, interpreter=interpreter, axis=1)
+
+        if config.prediction_cache > 0:
+            self.init_prediction_cache(config.prediction_cache)
+        data_source = DataSource.from_yaml(ds_path)
+        ddf_mapped = data_source.to_mapped_dataframe()
+        # this only makes really sense when we have a predict_batch_json method implemented ...
+        n_partitions = max(1, round(len(ddf_mapped) / config.batch_size))
+
+        # a persist is necessary here, otherwise it fails for n_partitions == 1
+        # the reason is that with only 1 partition we pass on a generator to predict_batch_json
+        ddf_mapped = ddf_mapped.repartition(npartitions=n_partitions).persist()
+        ddf_mapped_columns = ddf_mapped.columns
+
+        ddf_mapped["annotation"] = ddf_mapped[self.inputs_keys()].apply(
+            lambda x: sanitize(self.predict(**x.to_dict())), axis=1, meta=(None, object)
+        )
+
+        if config.interpret:
+            # TODO we should apply the same mechanism for the model predictions. Creating a new pipeline
+            #  for every partition
+            ddf_mapped["interpretations"] = ddf_mapped[
+                ddf_mapped_columns
+            ].map_partitions(_interpret_dataframe, pipeline=self, meta=(None, object))
+
+        ddf_source = data_source.to_dataframe()
+        ddf_source = ddf_source.repartition(npartitions=n_partitions).persist()
+
+        # We are sure that both data frames are aligned!
+        # A 100% safe way would be to set_index of both data frames on a meaningful column.
+        # The main problem are multiple csv files (read_csv("*.csv")), where the index starts from 0 for each file ...
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            ddf = dd.concat([ddf_source, ddf_mapped], axis=1)
+
+        # TODO @dcfidalgo we could calculate base metrics here (F1, recall & precision) using dataframe.
+        #  And include as part of explore metadata
+        #  Does it's simple???
+
+        ddf = DaskElasticClient(
+            host=es_config.es_host, retry_on_timeout=True, http_compress=True
+        ).save(ddf, index=es_config.es_index, doc_type=es_config.es_doc)
+
+        merged_metadata = {
+            **(config.metadata or {}),
+            "datasource": ds_path,
+            # TODO this should change when ui is normalized (action detail and action link naming)F
+            "explore_name": es_config.es_index,
+            "model": self.name,
+            "columns": ddf.columns.values.tolist(),
+        }
+
+        self.register_biome_prediction(
+            name=es_config.es_index, es_config=es_config, **merged_metadata
+        )
+        self.__prepare_es_index(es_config, force_delete=config.force_delete)
+        ddf = ddf.persist()
+        self._LOGGER.info(
+            "Data annotated successfully. You can explore your data here: %s",
+            f"{constants.EXPLORE_APP_ENDPOINT}/projects/default/explore/{es_config.es_index}",
+        )
+
+        return ddf
+
+    def serve(self, port: int, predictions: str):
+        if predictions:
+            self.init_prediction_logger(predictions)
+
+        app = server_simple.make_app(self, title=self.name)
+        CORS(app)
+
+        http_server = WSGIServer(("0.0.0.0", port), app)
+        self._LOGGER.info("Model loaded, serving on port %s", port)
+
+        http_server.serve_forever()
+
+    def register_biome_prediction(
+        self, name: str, es_config: ElasticsearchConfig, **kwargs
+    ):
+        """
+        Creates a new metadata entry for the incoming prediction
+
+        Parameters
+        ----------
+        name
+            A descriptive prediction name
+        pipeline
+            The pipeline used for the prediction batch
+        es_config:
+            The Elasticsearch configuration data
+        kwargs
+            Extra arguments passed as extra metadata info
+        """
+
+        metadata_index = constants.BIOME_METADATA_INDEX
+
+        es_config.client.indices.create(
+            index=metadata_index,
+            body={
+                "settings": {"index": {"number_of_shards": 1, "number_of_replicas": 0}}
+            },
+            params=dict(ignore=400),
+        )
+
+        parameters = {
+            **kwargs,
+            "pipeline": self.name,
+            "signature": self.inputs_keys() + [self.output()],
+            "predict_signature": self.inputs_keys(),
+            # TODO remove when ui is adapted
+            "inputs": self.inputs_keys(),  # backward compatibility
+        }
+
+        es_config.client.update(
+            index=metadata_index,
+            doc_type=es_config.es_doc,
+            id=es_config.es_index,
+            body={
+                "doc": dict(
+                    name=name, created_at=datetime.datetime.now(), **parameters
+                ),
+                "doc_as_upsert": True,
+            },
+        )
+
+    @staticmethod
+    def __prepare_es_index(es_config: ElasticsearchConfig, force_delete: bool):
+        dynamic_templates = [
+            {
+                data_type: {
+                    "match_mapping_type": data_type,
+                    "path_match": path_match,
+                    "mapping": {
+                        "type": "text",
+                        "fields": {"keyword": {"type": "keyword", "ignore_above": 256}},
+                    },
+                }
+            }
+            for data_type, path_match in [("*", "*.value"), ("string", "*")]
+        ]
+
+        if force_delete:
+            es_config.client.indices.delete(index=es_config.es_index, ignore=[400, 404])
+
+        es_config.client.indices.create(
+            index=es_config.es_index,
+            body={
+                "mappings": {es_config.es_doc: {"dynamic_templates": dynamic_templates}}
+            },
+            ignore=400,
+        )
 
     @staticmethod
     def __to_snake_case(name):
