+++ tests/text/pipelines/sequence_classifier_test.py
@@ -1,3 +1,4 @@
+import json
 import multiprocessing
 import os
 import tempfile
@@ -5,21 +6,19 @@ from time import sleep
 from typing import Optional
 
 import pandas as pd
-import json
 import pytest
 import requests
 import yaml
 from elasticsearch import Elasticsearch
+
 from biome.text import Pipeline
 from biome.text.commands.explore.explore import explore
 from biome.text.commands.serve.serve import serve
+from biome.text.defs import PipelineDefinition
 from biome.text.environment import ES_HOST
-from biome.text.pipelines.configuration import PipelineBuilder
 from biome.text.pipelines._impl.allennlp.classifier.pipeline import (
     AllenNlpTextClassifierPipeline,
 )
-from biome.text.pipelines._impl.allennlp.dataset_readers import SequenceClassifierReader
-from biome.text.pipelines._impl.allennlp.models import SequenceClassifier
 from biome.text.pipelines.sequence_classifier import SequenceClassifierPipeline
 from tests import DaskSupportTest
 from tests.test_context import TEST_RESOURCES
@@ -133,7 +132,6 @@ class SequenceClassifierTest(DaskSupportTest):
     validation_data = os.path.join(BASE_CONFIG_PATH, "validation.data.yml")
 
     def test_model_workflow(self):
-        self.check_train_other()
         self.check_train()
         # Check explore metadata override
         self.check_explore(
@@ -144,13 +142,14 @@ class SequenceClassifierTest(DaskSupportTest):
 
     def check_train_other(self):
         model_path = os.path.join(BASE_CONFIG_PATH, "model.new.yml")
-        classifier = AllenNlpTextClassifierPipeline(
-            model=SequenceClassifier, reader=SequenceClassifierReader
-        ).from_config(PipelineBuilder.from_file(model_path))
+        classifier = AllenNlpTextClassifierPipeline.from_config(
+            PipelineDefinition.from_file(model_path)
+        )
 
         classifier.learn(
             trainer=self.trainer_path, train=self.training_data, output=self.output_dir
         )
+
         classifier.learn(
             trainer=self.trainer_path,
             train=self.training_data,
